Improve reconstructed image quality

+++ jsolex-core/src/test/groovy/me/champeau/a4j/jsolex/processing/sun/workflow/JaggingCorrectionTest.groovy
@@ -29,8 +29,8 @@ class JaggingCorrectionTest extends Specification {
         def height = 1200
         def ellipse = createTestEllipse(width, height)
         def random = new Random(42)
-        def left = new int[height]
-        def right = new int[height]
+        def left = new double[height]
+        def right = new double[height]
         Arrays.fill(left, -1)
         Arrays.fill(right, -1)
         def data = new float[height][width]
@@ -43,7 +43,7 @@ class JaggingCorrectionTest extends Specification {
                 def shift = 3 * random.nextGaussian()
                 left[y] = (int) Math.round(x1 + shift)
                 right[y] = (int) Math.round(x2 + shift)
-                for (int x = Math.max(0, left[y]); x <= Math.min(width - 1, right[y]); x++) {
+                for (int x = (int) Math.max(0, left[y]); x <= (int) Math.min(width - 1, right[y]); x++) {
                     data[y][x] = 30000f
                 }
             }
@@ -98,8 +98,8 @@ class JaggingCorrectionTest extends Specification {
         def width = 800
         def height = 1200
         def ellipse = createTestEllipse(width, height)
-        def left = new int[height]
-        def right = new int[height]
+        def left = new double[height]
+        def right = new double[height]
         Arrays.fill(left, -1)
         Arrays.fill(right, -1)
